Rename `aviation.py` to `fleet.py` exit()
