+++ src/aviation/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ("passengers_per_day", "required_global_fleet")
 
-from aviation.aviation import passengers_per_day, required_global_fleet
+from aviation.fleet import passengers_per_day, required_global_fleet
